change dir in notification class.
