+++ app/backend/app/Services/Admins/Notifications/BaseSlackNotificationService.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace App\Services\Notifications;
+namespace App\Services\Admins\Notifications;
 
 use Illuminate\Notifications\Notifiable;
-use Illuminate\Support\Facades\Config;
 use App\Notifications\Admins\AdminUpdateNotification;
-use App\Services\Notifications\BaseSlackNotificationService;
+use Illuminate\Support\Facades\Config;
+use Illuminate\Notifications\Notification;
 
-class AdminsSlackNotificationService extends BaseSlackNotificationService
+class BaseSlackNotificationService
 {
+    use Notifiable;
+
     /**
      * send slack notification
      *
@@ -21,5 +23,16 @@ class AdminsSlackNotificationService extends BaseSlackNotificationService
         if (Config::get('app.env') !== 'testing') {
             $this->notify(new AdminUpdateNotification($message, $attachment));
         }
+    }
+
+    /**
+     * Route notifications for the Slack channel.
+     *
+     * @param  \Illuminate\Notifications\Notification  $notification
+     * @return string
+     */
+    public function routeNotificationForSlack($notification)
+    {
+        return Config::get('myapp.slack.url');
     }
 }
